:lock: Add refresh jwt token and authentication to routes

+++ src/http/routes/users/booklists/list entries/deleteListEntry.ts
@@ -1,6 +1,7 @@
 import { FastifyInstance } from 'fastify'
 import { prisma } from '../../../../../lib/prisma'
 import { z } from 'zod'
+import { authenticateAccess } from '../../../../../utils/authenticateJwt'
 
 const opts = {
   schema: {
@@ -16,6 +17,11 @@ const opts = {
 export async function deleteListEntry(app: FastifyInstance) {
   app.delete('/users/:username/book-lists/:book_list_id/list-entries/:entry_id', opts, 
   async (request, reply) => {
+    const requestUser = await authenticateAccess(request)
+    if(!requestUser) {
+      return reply.status(401).send({ message: 'Unauthorized request. Please log in first.' })
+    }
+
     const listEntryInformation = z.object({
       username: z.string().min(6),
       book_list_id: z.number().int(),
@@ -23,6 +29,7 @@ export async function deleteListEntry(app: FastifyInstance) {
     })
 
     const { username, book_list_id, entry_id } = listEntryInformation.parse(request.params)
+    if(username !== requestUser.username) return reply.status(403).send()
 
     const listEntry = await prisma.listEntry.delete({
       where: {
